move tests to correct package depth (output)
