+++ src/test/java/io/github/fruitcropxl/output/model/Fruit.java
